Extract Envoy helpers and centralize shared constants (#253)
