+++ pkg/translator/accesspolicy_test.go
@@ -447,9 +447,9 @@ func TestBuildRBACConfigWithCommonPolicies(t *testing.T) {
 
 	expectedPolicies := []string{
 		"custom-rule",
-		allowMCPSessionClosePolicyName,
-		allowAnyoneToInitializeAndListToolsPolicyName,
-		allowHTTPGet,
+		constants.AllowMCPSessionClosePolicyName,
+		constants.AllowAnyoneToInitializeAndListToolsPolicyName,
+		constants.AllowHTTPGetPolicyName,
 	}
 
 	for _, p := range expectedPolicies {
@@ -719,14 +719,14 @@ func TestTranslateAccessPolicyToRBAC(t *testing.T) {
 			expectedRules: map[string]expectedRule{
 				"rule-ext-auth": {
 					principal:      "spiffe://example.com/ns/default/sa/caller",
-					permissions:    []string{toolsCallMethod},
+					permissions:    []string{constants.ToolsCallMethod},
 					isExternalAuth: true,
 				},
 			},
 			expectedShadowRules: map[string]expectedRule{
 				"rule-ext-auth": {
 					principal:      "spiffe://example.com/ns/default/sa/caller",
-					permissions:    []string{toolsCallMethod},
+					permissions:    []string{constants.ToolsCallMethod},
 					isExternalAuth: true,
 				},
 			},
@@ -806,7 +806,7 @@ func TestBuildAllowAnyoneToInitializeAndListToolsPolicy(t *testing.T) {
 		t.Errorf("principal should be ANY")
 	}
 	matcher := policy.GetPermissions()[0].GetAndRules().GetRules()[0].GetSourcedMetadata().GetMetadataMatcher()
-	if matcher.GetFilter() != mcpProxyFilterName || matcher.GetPath()[0].GetKey() != "method" {
+	if matcher.GetFilter() != constants.MCPProxyFilterName || matcher.GetPath()[0].GetKey() != "method" {
 		t.Errorf("policy incorrectly configured")
 	}
 	orMethods := matcher.GetValue().GetOrMatch()
@@ -895,7 +895,7 @@ func verifyPolicy(t *testing.T, rbacPolicy *rbacconfigv3.Policy, expected expect
 			t.Fatal("expected NotRule for empty tools list")
 		}
 		methodRule := notRule.GetSourcedMetadata()
-		if methodRule == nil || methodRule.GetMetadataMatcher().GetValue().GetStringMatch().GetExact() != toolsCallMethod {
+		if methodRule == nil || methodRule.GetMetadataMatcher().GetValue().GetStringMatch().GetExact() != constants.ToolsCallMethod {
 			t.Errorf("expected 'disallow tools/call' permission for empty tools list")
 		}
 		return
@@ -907,7 +907,7 @@ func verifyPolicy(t *testing.T, rbacPolicy *rbacconfigv3.Policy, expected expect
 			return
 		}
 		methodRule := rbacPolicy.GetPermissions()[0].GetSourcedMetadata()
-		if methodRule == nil || methodRule.GetMetadataMatcher().GetValue().GetStringMatch().GetExact() != toolsCallMethod {
+		if methodRule == nil || methodRule.GetMetadataMatcher().GetValue().GetStringMatch().GetExact() != constants.ToolsCallMethod {
 			t.Errorf("expected tools/call method permission for external auth")
 		}
 		return
@@ -922,8 +922,8 @@ func verifyPolicy(t *testing.T, rbacPolicy *rbacconfigv3.Policy, expected expect
 
 	// Rule 1: method must be tools/call
 	methodRule := andRules.GetRules()[0].GetSourcedMetadata()
-	if methodRule == nil || methodRule.GetMetadataMatcher() == nil || methodRule.GetMetadataMatcher().GetFilter() != mcpProxyFilterName {
-		t.Errorf("first AND rule should be sourced metadata from %s", mcpProxyFilterName)
+	if methodRule == nil || methodRule.GetMetadataMatcher() == nil || methodRule.GetMetadataMatcher().GetFilter() != constants.MCPProxyFilterName {
+		t.Errorf("first AND rule should be sourced metadata from %s", constants.MCPProxyFilterName)
 		return
 	}
 	// Verify Path ["method"]
@@ -931,8 +931,8 @@ func verifyPolicy(t *testing.T, rbacPolicy *rbacconfigv3.Policy, expected expect
 		t.Errorf("tools/call matcher should have path ['method']")
 	}
 	// Verify Value "tools/call"
-	if methodRule.GetMetadataMatcher().GetValue().GetStringMatch().GetExact() != toolsCallMethod {
-		t.Errorf("tools/call matcher should match exact string %q", toolsCallMethod)
+	if methodRule.GetMetadataMatcher().GetValue().GetStringMatch().GetExact() != constants.ToolsCallMethod {
+		t.Errorf("tools/call matcher should match exact string %q", constants.ToolsCallMethod)
 	}
 
 	// Rule 2: tool names match
